Move AJV schema logic to its own file
